feat: remove react related
