--- example/App.css
@@ -1,7 +1,0 @@
-.App {
-  text-align: center;
-}
-#canvas {
-  width: 100vw;
-  height: 100vh;
-}
