+++ index.html
@@ -4,9 +4,14 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Vite App</title>
+  <style>
+    html, body {
+      margin: 0;padding: 0;
+    }
+  </style>
 </head>
 <body>
-  <div id="root"></div>
-  <script type="module" src="/example/main.jsx"></script>
+    <canvas id="canvas" style="width: 100vw; height: 100vh"></canvas>
+    <script type="module" src="/example/main.ts"></script>
 </body>
 </html>
